+++ src/three-d-secure/internal/bank-frame.js
@@ -5,6 +5,8 @@ var isVerifiedDomain = require("../../lib/is-verified-domain");
 var queryString = require("../../lib/querystring");
 var BraintreeError = require("../../lib/braintree-error");
 var sanitizeUrl = require("@braintree/sanitize-url").sanitizeUrl;
+var Client = require("../../client/client");
+var Promise = require("../../lib/promise");
 var errors = require("../shared/errors");
 var BUS_CONFIGURATION_REQUEST_EVENT =
   require("../../lib/constants").BUS_CONFIGURATION_REQUEST_EVENT;
@@ -22,29 +24,61 @@ module.exports = function () {
   bus.emit(BUS_CONFIGURATION_REQUEST_EVENT, handleConfiguration);
 };
 
+function isCardinalCommerce(client, configuration) {
+  return new Promise(function (resolve, reject) {
+    if (
+      /^https:\/\/([a-z0-9]+[.])*cardinalcommerce[.]com/.test(
+        configuration.acsUrl
+      )
+    ) {
+      return resolve(configuration.acsUrl);
+    }
+
+    return client
+      .request({
+        api: "clientApi",
+        method: "get",
+        endpoint: "payment_method_nonces/" + configuration.nonce,
+      })
+      .then(function (response) {
+        resolve(response.threeDSecureInfo.acsUrl);
+      })
+      .catch(function () {
+        reject(new BraintreeError(errors.THREEDS_LOOKUP_ERROR));
+      });
+  });
+}
+
 function handleConfiguration(configuration) {
   var input, field;
   var fields = { pareq: "PaReq", md: "MD", termUrl: "TermUrl" };
   var form = document.createElement("form");
+  var client = new Client(configuration.clientConfiguration);
 
   if (!isVerifiedDomain(configuration.termUrl)) {
     throw new BraintreeError(errors.THREEDS_TERM_URL_REQUIRES_BRAINTREE_DOMAIN);
   }
 
-  form.action = sanitizeUrl(configuration.acsUrl);
-  form.method = "POST";
+  return isCardinalCommerce(client, configuration)
+    .then(function (acsUrl) {
+      form.action = sanitizeUrl(acsUrl);
+      form.method = "POST";
 
-  for (field in fields) {
-    if (fields.hasOwnProperty(field)) {
-      input = document.createElement("input");
-      input.name = fields[field];
-      input.type = "hidden";
-      input.setAttribute("value", configuration[field]);
-      form.appendChild(input);
-    }
-  }
+      for (field in fields) {
+        if (fields.hasOwnProperty(field)) {
+          input = document.createElement("input");
+          input.name = fields[field];
+          input.type = "hidden";
+          input.setAttribute("value", configuration[field]);
+          form.appendChild(input);
+        }
+      }
 
-  document.body.appendChild(form);
+      document.body.appendChild(form);
 
-  form.submit();
+      form.submit();
+    })
+    .catch(function (err) {
+      Promise.reject(err);
+    });
 }